fix(storybook): use CommonJS config
